feat: ship the tsrx runtime from the rxfm package root
Move the runtime (render, RenderObservable, accumulate, interval, EMPTY) into
the library at src/lib/rxfm/runtime.ts and re-export it from the package root,
so consumers write `import { accumulate, interval } from "rxfm"` instead of a
relative `../runtime` path.

These helpers are useful with plain rxfm too: a shared/replaying derived value
(render/RenderObservable), a running fold (accumulate), and a reactive-period
clock (interval).

tsrx/runtime.ts becomes a thin shim re-exporting the library runtime via a
relative path, so it stays the module the transform injects `render` from and
resolves in every in-repo seam — including the transform's inference program,
which runs without tsconfig `paths`.
